requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-logging==3.6.0
google-cloud-aiplatform==1.28.0
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.2
google-cloud-secret-manager==2.16.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==11.0.0
pyparsing==3.1.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
fastapi==0.103.2
slack_bolt==1.18.0
uvicorn==0.23.2